Add macOS 26 support
- Also make the script executable and update README
